# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Language for the response (en = English, hi = Hindi)
 */
export type DiseaseDetectInputLanguage = typeof DiseaseDetectInputLanguage[keyof typeof DiseaseDetectInputLanguage];


export const DiseaseDetectInputLanguage = {
  en: 'en',
  hi: 'hi',
} as const;

export interface DiseaseDetectInput {
  /** Base64-encoded image data (without data URL prefix) */
  imageBase64: string;
  /** MIME type of the image (e.g. image/jpeg) */
  mimeType?: string;
  /** Language for the response (en = English, hi = Hindi) */
  language?: DiseaseDetectInputLanguage;
}

export interface TreatmentAdvice {
  immediate: string;
  chemical: string;
  organic: string;
  prevention: string;
}

export interface DiseaseResult {
  /** Name of the detected disease in English */
  diseaseName: string;
  /** Name of the detected disease in Hindi */
  diseaseNameHi: string;
  /** Whether the rice plant appears healthy */
  isHealthy: boolean;
  /** Confidence level (High / Medium / Low) */
  confidence: string;
  /**
     * Severity of the disease (Mild, Moderate, Severe)
     * @nullable
     */
  severity?: string | null;
  /** Brief description in English */
  description: string;
  /** Brief description in Hindi */
  descriptionHi: string;
  /** List of visible symptoms in English */
  symptoms: string[];
  /** List of visible symptoms in Hindi */
  symptomsHi: string[];
  treatment: TreatmentAdvice;
  treatmentHi: TreatmentAdvice;
  /**
     * Scientific species name from PlantNet identification
     * @nullable
     */
  plantNetSpecies?: string | null;
  /**
     * PlantNet identification confidence score (0-1)
     * @nullable
     */
  plantNetScore?: number | null;
}

export interface ErrorResponse {
  error: string;
}