# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: disease
    description: Rice disease detection
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /disease/detect:
    post:
      operationId: detectDisease
      tags: [disease]
      summary: Detect rice disease from image
      description: Accepts a base64-encoded rice leaf image and returns disease name, severity, and treatment advice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiseaseDetectInput"
      responses:
        "200":
          description: Detection result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiseaseResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    DiseaseDetectInput:
      type: object
      required:
        - imageBase64
      properties:
        imageBase64:
          type: string
          description: Base64-encoded image data (without data URL prefix)
        mimeType:
          type: string
          description: MIME type of the image (e.g. image/jpeg)
        language:
          type: string
          enum: [en, hi]
          description: Language for the response (en = English, hi = Hindi)
    TreatmentAdvice:
      type: object
      required:
        - immediate
        - chemical
        - organic
        - prevention
      properties:
        immediate:
          type: string
        chemical:
          type: string
        organic:
          type: string
        prevention:
          type: string
    DiseaseResult:
      type: object
      required:
        - diseaseName
        - diseaseNameHi
        - isHealthy
        - confidence
        - treatment
        - treatmentHi
        - description
        - descriptionHi
        - symptoms
        - symptomsHi
      properties:
        diseaseName:
          type: string
          description: Name of the detected disease in English
        diseaseNameHi:
          type: string
          description: Name of the detected disease in Hindi
        isHealthy:
          type: boolean
          description: Whether the rice plant appears healthy
        confidence:
          type: string
          description: Confidence level (High / Medium / Low)
        severity:
          type: ["string", "null"]
          description: Severity of the disease (Mild, Moderate, Severe)
        description:
          type: string
          description: Brief description in English
        descriptionHi:
          type: string
          description: Brief description in Hindi
        symptoms:
          type: array
          items:
            type: string
          description: List of visible symptoms in English
        symptomsHi:
          type: array
          items:
            type: string
          description: List of visible symptoms in Hindi
        treatment:
          $ref: "#/components/schemas/TreatmentAdvice"
        treatmentHi:
          $ref: "#/components/schemas/TreatmentAdvice"
        plantNetSpecies:
          type: ["string", "null"]
          description: Scientific species name from PlantNet identification
        plantNetScore:
          type: ["number", "null"]
          description: PlantNet identification confidence score (0-1)
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string