# lib/api-zod/src/generated/types/diseaseResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TreatmentAdvice } from './treatmentAdvice';

export interface DiseaseResult {
  /** Name of the detected disease in English */
  diseaseName: string;
  /** Name of the detected disease in Hindi */
  diseaseNameHi: string;
  /** Whether the rice plant appears healthy */
  isHealthy: boolean;
  /** Confidence level (High / Medium / Low) */
  confidence: string;
  /**
     * Severity of the disease (Mild, Moderate, Severe)
     * @nullable
     */
  severity?: string | null;
  /** Brief description in English */
  description: string;
  /** Brief description in Hindi */
  descriptionHi: string;
  /** List of visible symptoms in English */
  symptoms: string[];
  /** List of visible symptoms in Hindi */
  symptomsHi: string[];
  treatment: TreatmentAdvice;
  treatmentHi: TreatmentAdvice;
  /**
     * Scientific species name from PlantNet identification
     * @nullable
     */
  plantNetSpecies?: string | null;
  /**
     * PlantNet identification confidence score (0-1)
     * @nullable
     */
  plantNetScore?: number | null;
}